One more update to the zip-slip fix
Handle situations when removing the relative portions of path leads to an
absolute path. In such case we must make sure that the path is made relative so
that further `Path.Combine` doesn't result in us trying to create a file in the
root directory of the current drive/filesystem.